feat: move actions validator to dedicated folder
