+++ packages/datasource-customizer/src/collection-customizer.ts
@@ -17,7 +17,6 @@ import { OperatorDefinition } from './decorators/operators-emulate/types';
 import { RelationDefinition } from './decorators/relation/types';
 import { SearchDefinition } from './decorators/search/types';
 import { SegmentDefinition } from './decorators/segment/types';
-import ActionValidator from './decorators/validation/action';
 import { WriteDefinition } from './decorators/write/write-replace/types';
 import {
   ActionConfigurationValidationError,
@@ -27,6 +26,7 @@ import addExternalRelation from './plugins/add-external-relation';
 import importField from './plugins/import-field';
 import { TCollectionName, TColumnName, TFieldName, TSchema, TSortClause } from './templates';
 import { OneToManyEmbeddedDefinition, Plugin } from './types';
+import ActionValidator from './validators/action';
 
 export default class CollectionCustomizer<
   S extends TSchema = TSchema,
